[instancer/avar2] Progress
