+++ Lib/fontTools/varLib/instancer/__init__.py
@@ -452,10 +452,6 @@ class AxisLimits(_BaseAxisLimits):
             )
 
 
-        if usingAvar and "avar" in varfont:
-            avar = varfont["avar"]
-            normalizedLimits = avar.renormalizeAxisLimits(normalizedLimits, varfont)
-
         fvarAxes = fvar.getAxes()
         for tag, triple in normalizedLimits.items():
             minV, defaultV, maxV = fvarAxes[tag]
@@ -472,7 +468,13 @@ class AxisLimits(_BaseAxisLimits):
                 *triple, distanceNegative, distancePositive
             )
 
-        return NormalizedAxisLimits(normalizedLimits)
+        normalizedLimits = NormalizedAxisLimits(normalizedLimits)
+
+        if usingAvar and "avar" in varfont:
+            avar = varfont["avar"]
+            normalizedLimits = avar.renormalizeAxisLimits(normalizedLimits, varfont)
+
+        return normalizedLimits
 
 
 class NormalizedAxisLimits(_BaseAxisLimits):
